[ADD] 콘서트 구조 (Repository, Entity, Service, Dto(Response, Request) 생성)
