+++ src/main/java/com/hhplus/concert/common/BaseTimeEntity.java
@@ -1,0 +1,37 @@
+package com.hhplus.concert.common;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public class BaseTimeEntity {
+
+    @Column(name = "created_date")
+    @CreatedDate
+    private String createdDate;
+
+    @Column(name = "updated_date")
+    @LastModifiedDate
+    private String updatedDate;
+
+    /* 해당 엔티티를 저장하기 이전에 실행 */
+    @PrePersist
+    public void onPrePersist(){
+        this.createdDate = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy.MM.dd HH:mm"));
+        this.updatedDate = this.createdDate;
+    }
+
+    /* 해당 엔티티를 업데이트 하기 이전에 실행*/
+    @PreUpdate
+    public void onPreUpdate(){
+        this.updatedDate = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy.MM.dd HH:mm"));
+    }
+}
