+++ src/main/java/com/hhplus/concert/common/response/CommonResponse.java
@@ -1,0 +1,17 @@
+package com.hhplus.concert.common.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class CommonResponse<T> {
+
+    private String msg;
+    private T data;
+
+    @Builder
+    public CommonResponse(String msg, T data) {
+        this.msg = msg;
+        this.data = data;
+    }
+}
