+++ src/main/java/com/hhplus/concert/concert/Concert.java
@@ -1,0 +1,6 @@
+package com.hhplus.concert.concert;
+
+public record Concert (Long concertId, String concertName, String concertLocation,
+                       int fee) {
+
+}
