+++ src/main/java/com/hhplus/concert/concert/application/ConcertService.java
@@ -1,0 +1,7 @@
+package com.hhplus.concert.concert.application;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ConcertService {
+}
