+++ src/main/java/com/hhplus/concert/concert/infra/ConcertJpaRepository.java
@@ -1,0 +1,8 @@
+package com.hhplus.concert.concert.infra;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ConcertJpaRepository extends JpaRepository<ConcertEntity, Long> {
+}
