+++ src/main/java/com/hhplus/concert/concertItem/infra/ConcertItemEntity.java
@@ -1,0 +1,21 @@
+package com.hhplus.concert.concertItem.infra;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+
+import java.util.Date;
+
+@Entity
+@Table(name = "concert_item")
+@Getter
+public class ConcertItemEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long concertItemId;
+
+    private Date reservationStart;
+    private Date reservationEnd;
+    private Date concertStart;
+    private int seatId;
+}
