+++ src/main/java/com/hhplus/concert/concertItem/presentation/SearchSeatResponse.java
@@ -1,0 +1,16 @@
+package com.hhplus.concert.concertItem.presentation;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class SearchSeatResponse {
+
+    private List<Integer> seatAvailable;
+
+}
