+++ src/main/java/com/hhplus/concert/concertReservation/application/ConcertReservationService.java
@@ -1,0 +1,7 @@
+package com.hhplus.concert.concertReservation.application;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ConcertReservationService {
+}
