+++ src/main/java/com/hhplus/concert/concertReservation/infra/ConcertReservationJpaRepository.java
@@ -1,0 +1,6 @@
+package com.hhplus.concert.concertReservation.infra;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ConcertReservationJpaRepository extends JpaRepository<ConcertReservationEntity, Long> {
+}
