--- src/main/java/com/hhplus/concert/config/SwaggerConfig.java
@@ -1,25 +1,0 @@
-package com.hhplus.concert.config;
-
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .components(new Components())
-                .info(apiInfo());
-    }
-
-    private Info apiInfo() {
-        return new Info()
-                .title("항해 콘서트 명세서")
-                .description("항해 콘서트 API 명세서")
-                .version("1.0.0");
-    }
-}
