--- src/main/java/com/hhplus/concert/domain/Concert.java
@@ -1,6 +1,0 @@
-package com.hhplus.concert.domain;
-
-public record Concert (Long concertId, String concertName, String concertLocation,
-                       int fee) {
-
-}
