--- src/main/java/com/hhplus/concert/domain/User.java
@@ -1,4 +1,0 @@
-package com.hhplus.concert.domain;
-
-public record User(Long userId, String userName) {
-}
