--- src/main/java/com/hhplus/concert/dto/ChargeRequest.java
@@ -1,14 +1,0 @@
-package com.hhplus.concert.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class ChargeRequest {
-
-    private Long userId;
-    private String amount;
-    private String token;
-
-}
