--- src/main/java/com/hhplus/concert/dto/PaymentRequest.java
@@ -1,14 +1,0 @@
-package com.hhplus.concert.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class PaymentRequest {
-
-    private Long userId;
-    private int price;
-    private String token;
-
-}
