--- src/main/java/com/hhplus/concert/dto/SearchSeatResponse.java
@@ -1,16 +1,0 @@
-package com.hhplus.concert.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class SearchSeatResponse {
-
-    private List<Integer> seatAvailable;
-
-}
