--- src/main/java/com/hhplus/concert/dto/TokenRequest.java
@@ -1,13 +1,0 @@
-package com.hhplus.concert.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class TokenRequest {
-
-    private Long uuid;
-    private Long waistId;
-
-}
