--- src/main/java/com/hhplus/concert/dto/TokenResponse.java
@@ -1,17 +1,0 @@
-package com.hhplus.concert.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class TokenResponse {
-
-    private String token;
-    private LocalDateTime expires;
-
-}
