--- src/main/java/com/hhplus/concert/dto/UserBalanceResponse.java
@@ -1,15 +1,0 @@
-package com.hhplus.concert.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class UserBalanceResponse {
-
-    private int balance;
-
-}
