+++ src/main/java/com/hhplus/concert/seat/application/SeatService.java
@@ -1,0 +1,7 @@
+package com.hhplus.concert.seat.application;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class SeatService {
+}
