+++ src/main/java/com/hhplus/concert/seat/infra/SeatJpaRepository.java
@@ -1,0 +1,8 @@
+package com.hhplus.concert.seat.infra;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SeatJpaRepository extends JpaRepository<SeatEntity, Long> {
+}
