+++ src/main/java/com/hhplus/concert/token/presentation/TokenRequest.java
@@ -1,0 +1,13 @@
+package com.hhplus.concert.token.presentation;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class TokenRequest {
+
+    private Long uuid;
+    private Long waistId;
+
+}
