+++ src/main/java/com/hhplus/concert/token/presentation/TokenResponse.java
@@ -1,0 +1,17 @@
+package com.hhplus.concert.token.presentation;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class TokenResponse {
+
+    private String token;
+    private LocalDateTime expires;
+
+}
