+++ src/main/java/com/hhplus/concert/user/application/User.java
@@ -1,0 +1,4 @@
+package com.hhplus.concert.user.application;
+
+public record User(Long userId, String userName) {
+}
