+++ src/main/java/com/hhplus/concert/user/application/UserService.java
@@ -1,0 +1,4 @@
+package com.hhplus.concert.user.application;
+
+public class UserService {
+}
