+++ src/main/java/com/hhplus/concert/user/infra/UserEntity.java
@@ -1,0 +1,16 @@
+package com.hhplus.concert.user.infra;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+
+@Entity
+@Table(name = "user")
+@Getter
+public class UserEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long userId;
+    private String userName;
+
+}
