+++ src/main/java/com/hhplus/concert/user/infra/UserJpaRepository.java
@@ -1,0 +1,8 @@
+package com.hhplus.concert.user.infra;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserJpaRepository extends JpaRepository<UserEntity, Long> {
+}
