+++ src/main/java/com/hhplus/concert/userCharge/application/UserChargeService.java
@@ -1,0 +1,7 @@
+package com.hhplus.concert.userCharge.application;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserChargeService {
+}
