+++ src/main/java/com/hhplus/concert/userCharge/presentation/UserChargeResponse.java
@@ -1,0 +1,15 @@
+package com.hhplus.concert.userCharge.presentation;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class UserChargeResponse {
+
+    private int balance;
+
+}
