+++ src/main/java/com/hhplus/concert/waitingQue/application/WaitingQueService.java
@@ -1,0 +1,7 @@
+package com.hhplus.concert.waitingQue.application;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class WaitingQueService {
+}
