+++ src/main/java/com/hhplus/concert/waitingQue/presentation/WaitingQueResponse.java
@@ -1,0 +1,15 @@
+package com.hhplus.concert.waitingQue.presentation;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class WaitingQueResponse {
+
+    private Long userId;
+    private int waitingOrder;
+
+}
